Add Tales of Graces (STGJAF)
Context: The Spanish fan translation that released a long time ago was uh, lacking... in some places. I made an attempt to finish up the missing textures and ran into a problem with the initial (controller strap) screen using an unknown compression, one of these codes modifies the file in mem2 so the decompression is taken advantage of and create a black screen, then it forces a 'A button press' to quickly skip the screen. Boom! no more Japanese strap screen, problem solved.