Add Sonic Heroes (G9SE8P)
My pride and joy! The metallic costumes outside of multiplayer. It took so long to get right, the bool used to enable the costumes only affected fly/power types, I had to improvise with the speed type by taking advantage of loading times to manipulate the 2P paths at the correct time and revert them as needed.