Add Sonic Riders (GXEE8P)
Once upon a time, the perfect plan to make a Sonic tool to rule them all... The Chao Transfer Tool. It is a way to send chao back and forth SADX and SA2B, but what happens when you get ambitious? You begin adding stuff you don't need to add, like a "Ring Giver" mode, SR was a target game for this, unfortunately the savefile's checksum prevented me from making this possible, so this code here would "solve" that. Except, it was annoying to rely on a code to get this feature working, so in the end I cracked the checksum, it's crc16 with a 0xFFFF initial value and inverted result.