SADX codes for using DLC music in-game
SADX was the first game I tried making a pointer code, the plan was to make Metal Sonic play differently compared to Sonic, so I would use this code to make MS slower so his top speed animation would never play, this did not work correctly, the action stages are mostly unplayable and the animation still plays for a bit before the code takes effect.

In the end I settled for making a code so that if MS is selected he cannot collect rings. I recently found out that 4 DLC music tracks were kept in the disc, so I made some codes to load them into the main game.

The codes for setting the Time of Day may become "permanent" if you save, you can always load a different time and save again to revert it.