Added more Sonic Mega Collection codes
NOTE: Sonic 1 "JP2" is in the game disc, but upon loading it via the official combo (zzzz,up,down,etc) the game reverts to the normal ROM. The new code writes the changes to the normal ROM and updates the checksum, this requires two versions, the standard one and one for the different checksum caused by the "Show Press Start" code. A check was added so that all combinations work at the same time.

Sonic 1 "JP2" is the default ROM with the "spike bug" restored. I don't know what the intention was, perhaps to make the game harder without giving up the other updates? Maybe that's why it doesn't work, they realized it didn't offer anything interesting.