Removed one of the text-expanding cheats, since it's no longer needed.
The line in question "Stop the lesson and return to the selection screen?" originally overwrote 00s that the game uses, causing a crash in the tutorial mode on Wii, and warnings on Dolphin. The cheat itself wasn't the problem but the actual changes to the dol. Instead, the line is now "Stop the lesson and return to the selection?"
Added some simple codes for controlling vfilter values and VI scale/position.