Fix main.dol crashing in the tutorial mode on Wii for Haruhi Gekidou
I ignored this issue for a long time and thought the problem came from incorrectly editing the tutorial scripts. In reality, I overwrote some 00s in the dol that are used by the tutorial code. Dolphin wouldn't crash, it just gave me a few invalid read warnings and continued running.