Riivolution patch for Mickey and the Magic Brush
This allows playing the (improved) JP release of Epic Mickey in any of the supported languages. By default it will start in English, using other languages should work, but note that the opening, credits, and possibly other scenes will be mute, this can be worked around by using a bik muxer to duplicate an existing audio track.

The strap screen is skipped by cmdline.txt, edit it to restore it. The Home Menu needs to be manually edited to other languages, English is used.