Finished drawing font graphics
From D to z with accented vowels, cedilla, inverted question/exclamation, ü, commas, etc.

Just apply the bin to ROM address 0x6328B8 and then you'll have access to a halfwidth font in the game that has all needed characters for English, Spanish, and possibly others like Portuguese. This fills up nearly all space for new characters, more space could be used since the game maps it that way but there's data in that area of the ROM that certainly shouldn't be overwritten.